# docs/swagger.yaml
definitions:
...
  api.getAllPetugasKesehatanResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.petugasKesehatanResponse'
        type: array
      total:
        type: integer
    type: object
...
  api.insertPetugasKesehatanRequest:
    properties:
      email:
        type: string
      id_skpd:
        type: string
      nama:
        type: string
      password:
        type: string
    type: object
  api.insertPetugasKesehatanResponse:
    properties:
      id:
        type: string
    type: object
...
  api.petugasKesehatanResponse:
    properties:
      created_date:
        type: string
      email:
        type: string
      id:
        type: string
      id_pengguna:
        type: string
      id_skpd:
        type: string
      intervensi_count:
        description: jumlah intervensi terkait
        type: integer
      jenis_skpd:
        type: string
      nama:
        type: string
      skpd:
        type: string
      updated_date:
        type: string
    type: object
...
  object.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing stunting data
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stunting Web API
  version: 0.0.2
paths:
...
  /api/admin/petugas-kesehatan/get:
    get:
      consumes:
      - application/json
      description: |-
        Get petugas kesehatan data based on query parameter (Admin only)

        Response data varies by parameter:
        - Without id parameter: Returns all petugas kesehatan with total count
        - With id parameter: Returns specific petugas kesehatan data

        Petugas kesehatan data includes: nama, email, SKPD info, intervensi count, creation/update dates
      parameters:
      - description: Petugas Kesehatan ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Petugas kesehatan data retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.getAllPetugasKesehatanResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Petugas kesehatan not found
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Get petugas kesehatan data
      tags:
      - admin
  /api/admin/petugas-kesehatan/insert:
    post:
      consumes:
      - application/json
      description: |-
        Insert new petugas kesehatan data (Admin only)

        Creates a complete petugas kesehatan record including:
        - Creates pengguna account with email and hashed password
        - Creates petugas_kesehatan record linked to SKPD
        - Sets role as "petugas kesehatan" in pengguna table
        - Validates email uniqueness and SKPD existence
      parameters:
      - description: Petugas kesehatan data
        in: body
        name: petugas
        required: true
        schema:
          $ref: '#/definitions/api.insertPetugasKesehatanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Petugas kesehatan inserted successfully
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.insertPetugasKesehatanResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Insert new petugas kesehatan
      tags:
      - admin
...
swagger: "2.0"